% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eClustaLSEC_GEM.R
\name{eClustaLSEC_GEM}
\alias{eClustaLSEC_GEM}
\title{Automated edge clustering using a finite mixture prior}
\usage{
eClustaLSEC_GEM(
  A,
  K = NULL,
  p = NULL,
  maxIter = 1000,
  maxIterVB = 100,
  eps = 1e-04,
  QNSteps = 25,
  CGSteps = 25,
  UV_init_sd = 1,
  alpha_init = NULL,
  a_u = 1,
  b_u = 1,
  a_s = 1,
  b_s = 1,
  a_v = a_u,
  b_v = b_u,
  a_r = a_s,
  b_r = b_s,
  a_a = 1,
  b_a = 200
)
}
\arguments{
\item{A}{igraph object}

\item{K}{integer. Maximum number of clusters}

\item{p}{integer. Number of latent dimensions}

\item{maxIter}{integer. Maximum number of iterations of the GEM algorithm}

\item{maxIterVB}{integer. Maximum number of iterations of the VB algorithm for approximating E step}

\item{eps}{numeric. Convergence threshold for relative change of ELBO in variational E step}

\item{QNSteps}{integer.  Maximum number of quasi-Newton steps during each iteration for W}

\item{CGSteps}{integer.  Maximum number of conjugate gradient steps during each
iteration for U and V}

\item{UV_init_sd}{numeric. SD used to randomly initialize variance of U and V.}

\item{alpha_init}{numeric. Initial value of alpha. If NULL then set to be the prior mean (a_a/b_a).}

\item{a_u, }{b_u, a_s, b_s, a_v, b_v, a_r, b_r, a_a, b_a numeric.  Hyper-parameters.}
}
\value{
Object of class 'eClustaLSEC_GEM'.
\itemize{
\item estimates a list providing the MAP estimates of the model parameters
\item algoParms a list providing the optimization parameters
\item userInputs a list providing the number of clusters, dimension of the latent space, and the hyperparameters supplied by the user
}
}
\description{
'eClustaLSEC_GEM' implements the variational Bayes generalized EM algorithm for clustering edges of a
network using a latent space approach. Use multiple starting points to find the best mode!
}
