% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICL.eClustaLSEC_GEM.R
\name{ICL.eClustaLSEC_GEM}
\alias{ICL.eClustaLSEC_GEM}
\title{ICL for edge clustering model using a finite mixture prior}
\usage{
ICL.eClustaLSEC_GEM(eCl, A, ...)
}
\arguments{
\item{eCl}{object of class 'eClustaLSEC_GEM'}

\item{A}{igraph object}

\item{...}{Further parameters to use in \code{optim()}.}
}
\description{
Compute the Integrated Complete Likelihood for edge clustering object
}
\details{
Higher is better.
}
