% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eClustaLSEC_Sim.R
\name{eClustaLSEC_Sim}
\alias{eClustaLSEC_Sim}
\title{Simulate a network according to the latent space model}
\usage{
eClustaLSEC_Sim(
  p = 2,
  K = 2 * p,
  n = 400,
  M = round(n * (n - 1) * 0.05),
  conc = c(1.5, 2.5, 6, 25, 15, 40, 100, 1.75, 5, 10, 50, 10, 15, 30)[7 * (p - 2) + K -
    1],
  UVShape = c(20, 60)[(K > 5) + 1],
  UVRate = 4
)
}
\arguments{
\item{p}{integer. Number of latent dimensions}

\item{K}{integer. True number of clusters}

\item{n}{integer. Number of actors in the network}

\item{M}{integer. Number of edges in the network}

\item{conc}{concentration parameter of the mixtures of von Mises-Fisher distributions to generate U, V}

\item{UVShape}{shape parameter of the Gamma distribution to generate the magnitude of U, V}

\item{UVRate}{rate parameter of the Gamma distribution to generate the magnitude of U, V}
}
\description{
Simulate a network according to the latent space model
}
